% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armengol.R
\docType{data}
\name{armengol}
\alias{armengol}
\title{CNV data from Armengol, et al paper}
\format{A data frame with 261 rows and 121 variables:
\describe{
  \item{pop}{human major group (CEU, YRI or CHB+JPT)}
  \item{HIRA.x}{number of repetitions for the 'HIRA.x' CNV}
  ...
}}
\usage{
data(armengol)
}
\description{
A data frame with 261 observations containing information about 120
CNVs for three human major groups (CEU, YRI and CHB+JPT)
}
\references{
Armengol et al. Identification of Copy Number Variants Defining Genomic Differences
among Major Human Groups. Plos One, 2009, e7230+.
}
\keyword{datasets}
