% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesOmicAssocByChr.R
\name{bayesOmicAssocByChr}
\alias{bayesOmicAssocByChr}
\title{Function that performs the bayesOmic analysis by chromosome of a RangedSummarizedExperiment}
\usage{
bayesOmicAssocByChr(group, data, chrList)
}
\arguments{
\item{group}{name of grouping variable}

\item{data}{omic data. It must be a RangedSummarizedExperiment object}

\item{chrList}{vector list of all seqnames to analyze}
}
\description{
Function that performs the bayesOmic analysis by chromosome of a RangedSummarizedExperiment
}
