% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSNPassoc.R
\name{bayesSNPassoc}
\alias{bayesSNPassoc}
\title{Bayesian model to analyze SNP data}
\usage{
bayesSNPassoc(group, data, sep.allele = "", annotation, chr,
  call.rate = 0.9, min.freq = 0.05, sig.level = 0.05,
  n.iter.burn.in = 1000, n.iter = 5000, thin = 10, n.chain = 2,
  ...)
}
\arguments{
\item{group}{vector with one value per individual, stores information about group belonging to (Case/Control, PopA/PopB/PopC...)}

\item{data}{matrix with the genotype of each SNP (columns) corresponding to each individual (rows)}

\item{sep.allele}{how alleles are separeted. Argument to be passed through SNPassoc::setupSNP() when SNP data is in a matrix or data.frame. Default is "".}

\item{annotation}{}

\item{chr}{}

\item{call.rate}{}

\item{min.freq}{}

\item{sig.level}{significance level}

\item{n.iter.burn.in}{}

\item{n.iter}{}

\item{thin}{}

\item{n.chain}{}

\item{other}{arguments to be passed trough SNPassoc::setupSNP}
}
\description{
Bayesian model to analyze SNP data
}
