% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGenotypeData.R
\name{createGenotypeData}
\alias{createGenotypeData}
\title{Creates a GenotyeData object with genetic and phenotypic data}
\usage{
createGenotypeData(geno, pheno, scanIDcol = 1)
}
\arguments{
\item{geno}{object of class 'GdsGenotypeReader' (e.g., PLINK in GDS format)}

\item{pheno}{a data.frame with phenotypic data}

\item{scanIDcol}{the column having the identification number in the pheno data.frame (must be the same as in the PLINK data)}
}
\description{
Creates a GenotyeData object with genetic and phenotypic data
}
