% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoBayesCNVassoc.R
\name{getInfoBayesCNVassoc}
\alias{getInfoBayesCNVassoc}
\title{getInfoBayesCNVassoc}
\usage{
getInfoBayesCNVassoc(x, N.groups, N.cnvs, names.groups, names.CNVs, quantiles)
}
\arguments{
\item{x}{}

\item{N.groups}{}

\item{N.cnvs}{}

\item{names.groups}{}

\item{names.CNVs}{}

\item{quantiles}{}
}
\description{
getInfoBayesCNVassoc
}
\keyword{internal}
