% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoBayesSNPassoc.R
\name{getInfoBayesSNPassoc}
\alias{getInfoBayesSNPassoc}
\title{getInfoBayesSNPassoc}
\usage{
getInfoBayesSNPassoc(x, Ngroups, Nvar, names.groups, names.SNPs)
}
\arguments{
\item{x}{}

\item{Ngroups}{}

\item{Nvar}{}

\item{names.groups}{}

\item{names.SNPs}{}
}
\description{
getInfoBayesSNPassoc
}
\keyword{internal}
