% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\alias{print.parameters}
\title{Get Parameters for a Bayesian model (SNP or CNV)}
\usage{
\method{print}{parameters}(x)
}
\arguments{
\item{x}{An object of class 'bayesSNPassoc' or 'bayesCNVassoc'}
}
\description{
Get Parameters for a Bayesian model (SNP or CNV)
}
