% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Creates a Heatmap using coefficients of Bayesian model}
\usage{
makeHeatmap(x, quantiles = c(0.2, 0.5, 0.6, 0.8), ...)
}
\arguments{
\item{x}{object of class 'bayesSNPassoc'}

\item{quantiles}{}
}
\description{
Creates a Heatmap using coefficients of Bayesian model
}
