% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\docType{data}
\name{mod}
\alias{mod}
\title{Fitted model for simulated data}
\format{An object of class \code{bayesSNPassoc} of length 7.}
\usage{
data(mod)
}
\description{
Fitted model for simulated data
}
\keyword{datasets}
