% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayesCNVassoc.R
\name{plot.bayesCNVassoc}
\alias{plot.bayesCNVassoc}
\title{plot.bayesCNVassoc}
\usage{
plot.bayesCNVassoc(x, mfrow, pos.legend = "bottomright", ...)
}
\arguments{
\item{x}{object of class 'bayesCNVassoc'}

\item{mfrow}{}

\item{pos.legend}{position of the plot legend}
}
\description{
plot.bayesCNVassoc
}
