% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayesOmic.R
\name{plot.bayesOmic}
\alias{plot.bayesOmic}
\title{plot.bayesOmicAssoc}
\usage{
plot.bayesOmic(x, type = "specific", show.y.lab = FALSE, ...)
}
\arguments{
\item{x}{object of class 'bayesOmic'}

\item{show.y.lab}{wether to show or not y labs}

\item{mfrow}{}

\item{pos.legend}{position of the plot legend}
}
\description{
plot.bayesOmicAssoc
}
