% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayesSNPassoc.R
\name{plot.bayesSNPassoc}
\alias{plot.bayesSNPassoc}
\title{plot.bayesSNPassoc}
\usage{
plot.bayesSNPassoc(x, type = "specific", mfrow, ...)
}
\arguments{
\item{x}{object of class 'bayesSNPassoc'}

\item{type}{}

\item{mfrow}{}
}
\description{
plot.bayesSNPassoc
}
\keyword{internal}
