% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bayesCNVassoc.R
\name{print.bayesCNVassoc}
\alias{print.bayesCNVassoc}
\title{print.bayesCNVassoc}
\usage{
\method{print}{bayesCNVassoc}(x, sig.CNVs = c(0.05, 0.95), ...)
}
\arguments{
\item{x}{object of class 'bayesCNVassoc'}

\item{sig.CNVs}{}
}
\description{
print.bayesCNVassoc
}
\keyword{internal}
