% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bayesOmic.R
\name{print.bayesOmic}
\alias{print.bayesOmic}
\title{print.bayesOmicAssoc}
\usage{
\method{print}{bayesOmic}(x, ...)
}
\arguments{
\item{x}{object of class 'bayesOmic'}

\item{...}{}

\item{x}{}
}
\description{
print.bayesOmicAssoc
}
