% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.data.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated SNP data}
\format{A data frame with 800 observations on the following 73 variables.  
\describe{
  \item{caco}{Case or Control}
  \item{snp74519580}{genotype for SNP 74519580 (AA/AB/BB)}
  ...
}}
\usage{
data(sim.data)
}
\description{
A data frame with 800 observations (cases and controls) containing information about 72
SNPs
}
\keyword{datasets}
