% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFileGenoDat.R
\name{writeFileGenoDat}
\alias{writeFileGenoDat}
\title{writeFileGenoDat}
\usage{
writeFileGenoDat(gg, allesum = FALSE, caco)
}
\arguments{
\item{gg}{}

\item{allesum}{}

\item{caco}{}
}
\description{
writeFileGenoDat
}
\keyword{internal}
