% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCNVassoc.R
\name{bayesCNVassoc}
\alias{bayesCNVassoc}
\alias{plot.bayesCNVassoc}
\alias{print.bayesCNVassoc}
\title{Bayesian model to analyze CNV data}
\usage{
bayesCNVassoc(
  cnvs,
  y,
  QC = 0.9,
  method = "JAGS",
  n.iter.burn.in = 10000,
  n.iter = 30000,
  thin = 50,
  n.chain = 2,
  alpha.corrected,
  numint.maxfeval = 10000,
  ...
)
}
\arguments{
\item{cnvs}{cnvs}

\item{y}{y}

\item{QC}{}

\item{method}{method can be 'JAGS' or 'INLA'}

\item{n.iter.burn.in}{}

\item{n.iter}{}

\item{thin}{}

\item{n.chain}{}
}
\description{
Bayesian model to analyze CNV data
}
