% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesOmicAssoc.R
\name{bayesOmicAssoc}
\alias{bayesOmicAssoc}
\alias{plot.bayesOmic}
\alias{print.bayesOmic}
\title{Bayesian model to analyze CNV data}
\usage{
bayesOmicAssoc(group, data, miss.indiv = 0.9, sig.level = 0.05,
  n.iter.burn.in = 1000, n.iter = 5000, thin = 10, n.chain = 2,
  ...)
}
\arguments{
\item{group}{}

\item{data}{}

\item{miss.indiv}{minimum percentage of missing individuals allowed. Default is 0.9}

\item{sig.level}{significance level}

\item{n.iter.burn.in}{}
}
\description{
Bayesian model to analyze CNV data
}
