% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSNPassoc.R
\name{bayesSNPassoc}
\alias{bayesSNPassoc}
\title{Bayesian model to analyze SNP data}
\usage{
bayesSNPassoc(formula, data, chr, call.rate = 0.9, min.freq = 0.05,
  sig.level = 0.05, method = "JAGS", n.iter.burn.in = 1000,
  n.adapt = 100, n.iter = 5000, thin = 5, n.chain = 2,
  n.cores = 1, ...)
}
\arguments{
\item{data}{setupSNP data.frame (SNPassoc) with genotypes of each SNP (columns) corresponding to each individual (rows) and phenotypic variables.}

\item{chr}{}

\item{call.rate}{}

\item{min.freq}{}

\item{sig.level}{significance level.}

\item{method}{estimating method. It can be INLA (default) or JAGS.}

\item{n.iter.burn.in}{}

\item{n.iter}{}

\item{thin}{}

\item{n.chain}{}

\item{...}{other arguments to be passed trough INLA::inla}
}
\description{
Bayesian model to analyze SNP data
}
