% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSNPassoc.R
\name{bayesSNPassoc}
\alias{bayesSNPassoc}
\alias{plot.bayesSNPassoc}
\alias{print.bayesSNPassoc}
\title{Bayesian model to analyze SNP data}
\usage{
bayesSNPassoc(
  snps,
  y,
  annotation,
  chr,
  QC = 0.9,
  min.freq = 0.05,
  method = "JAGS",
  n.iter.burn.in = 10000,
  n.iter = 30000,
  thin = 50,
  n.chain = 2,
  ...
)
}
\arguments{
\item{snps}{matrix with the phenotype for each SNP (columns) corresponding to each individual (rows)}

\item{y}{vector with one value per individual, stores information about group belonging (Case/Control, PopA/PopB/PopC...)}

\item{annotation}{}

\item{chr}{}

\item{QC}{}

\item{min.freq}{}

\item{method}{method can be 'JAGS' or 'INLA'}

\item{n.iter.burn.in}{}

\item{n.iter}{}

\item{thin}{}

\item{n.chain}{}
}
\description{
Bayesian model to analyze SNP data
}
