% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesSNPassoc.R
\name{bayesSNPassoc}
\alias{bayesSNPassoc}
\title{Bayesian model to analyze SNP data}
\usage{
bayesSNPassoc(snps, y, annotation, chr, call.rate = 0.9,
  min.freq = 0.05, method = "JAGS", n.iter.burn.in = 10000,
  n.iter = 30000, thin = 50, n.chain = 2, ...)
}
\arguments{
\item{snps}{matrix with the genotype of each SNP (columns) corresponding to each individual (rows)}

\item{y}{vector with one value per individual, stores information about group belonging to (Case/Control, PopA/PopB/PopC...)}

\item{annotation}{}

\item{chr}{}

\item{call.rate}{}

\item{min.freq}{}

\item{method}{method can be 'JAGS' or 'INLA'}

\item{n.iter.burn.in}{}

\item{n.iter}{}

\item{thin}{}

\item{n.chain}{}
}
\description{
Bayesian model to analyze SNP data
}
