% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConvergence.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Check model convergence of coefficients from the Bayesian model}
\usage{
checkConvergence(x, type = "Markov chain", parameter = "alpha")
}
\arguments{
\item{x}{object of class 'bayesOmic'}

\item{parameter}{}
}
\description{
Check model convergence of coefficients from the Bayesian model
}
