% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConvergence.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Function to check convergence of Bayesian model}
\usage{
checkConvergence(x, type = "Markov chain", parameter = "alpha", ...)
}
\arguments{
\item{x}{}

\item{type}{}

\item{parameter}{}
}
\description{
Function to check convergence of Bayesian model
}
