% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayesOmic.R
\name{plot.bayesOmic}
\alias{plot.bayesOmic}
\title{plot.bayesOmicAssoc}
\usage{
\method{plot}{bayesOmic}(x, type = "specific", ...)
}
\arguments{
\item{x}{object of class 'bayesOmic'}

\item{mfrow}{}

\item{pos.legend}{position of the plot legend}
}
\description{
plot.bayesOmicAssoc
}
